The src file is missing. Fixed a read only problem
The source file was missing and even if trying to make the one under production source it was still read only file without option to be changed. The solution is to create a folder src, make it source from project structure, and create the files. the main file is created with .java extension.